+++ R/get_power.R
@@ -16,8 +16,8 @@
 #'   values are \dQuote{DAILY}, \dQuote{INTERANNUAL} and \dQuote{CLIMATOLOGY}.
 #'   See argument details for more.
 #' @param lonlat A numeric vector of geographic coordinates for a cell or region
-#'   entered as x, y coordinates.  Not used when \code{temporal_average} is set
-#'   to \dQuote{CLIMATOLOGY}.  See argument details for more.
+#'   entered as x, y coordinates or \dQuote{GLOBAL} for global coverage (only
+#'   used for \dQuote{CLIMATOLOGY}).  See argument details for more.
 #' @param dates A character vector of start and end dates in that order,\cr
 #'   \emph{e.g.}, \code{dates = c("1983-01-01", "2017-12-31")}.
 #'   Not used when\cr \code{temporal_average} is set to \dQuote{CLIMATOLOGY}.
@@ -63,6 +63,10 @@
 #'  given region, \emph{e.g.}, a bounding box for the southwestern corner of
 #'  Australia: \code{lonlat = c(112.5, -55.5, 115.5, -50.5)}. *Maximum area
 #'  processed is 4.5 x 4.5 degrees (100 points).}
+#'
+#'  \item{For global coverage}{To get global coverage for CLIMATOLOGY, supply
+#'  \dQuote{GLOBAL} while also specifying \dQuote{CLIMATOLOGY} for the
+#'  \code{temporal_average}.}
 #' }
 #'
 #' @section Argument details for \code{dates}: If one date only is provided, it
@@ -91,21 +95,32 @@
 #' # humidity and precipitation for January 1 1985
 #' # for Kingsthorpe, Queensland, Australia
 #' ag_d <- get_power(
-#'    community = "AG",
-#'    lonlat = c(151.81, -27.48),
-#'    pars = c("RH2M", "T2M", "PRECTOT"),
-#'    dates = "1985-01-01",
-#'    temporal_average = "DAILY"
+#'   community = "AG",
+#'   lonlat = c(151.81, -27.48),
+#'   pars = c("RH2M", "T2M", "PRECTOT"),
+#'   dates = "1985-01-01",
+#'   temporal_average = "DAILY"
 #' )
 #'
-#' ag_d
+#' # Fetch single point climatology for air temperature
+#' ag_c_point <- get_power(
+#'   community = "AG",
+#'   pars = "T2M",
+#'   c(151.81, -27.48),
+#'   temporal_average = "CLIMATOLOGY"
+#' )
 #'
-#' # Fetch global AG climatology for temperature
-#' ag_c <- get_power(community = "AG",
-#'    pars = "T2M",
-#'    temporal_average = "CLIMATOLOGY")
+#' ag_c_point
 #'
-#' ag_c
+#' # Fetch global AG climatology for air temperature
+#' ag_c_global <- get_power(
+#'   community = "AG",
+#'   pars = "T2M",
+#'   lonlat = "GLOBAL",
+#'   temporal_average = "CLIMATOLOGY"
+#' )
+#'
+#' ag_c_global
 #'
 #' # Fetch interannual solar cooking parameters
 #' # for a given region
@@ -127,7 +142,7 @@
 get_power <- function(community,
                       pars,
                       temporal_average,
-                      lonlat = NULL,
+                      lonlat,
                       dates = NULL) {
   if (is.character(temporal_average)) {
     temporal_average <- toupper(temporal_average)
@@ -144,37 +159,50 @@ get_power <- function(community,
   if (is.character(pars)) {
     pars <- toupper(pars)
   }
-  if (is.character(community)) {
-    community <- toupper(community)
+  if (is.character(lonlat)) {
+    lonlat <- toupper(lonlat)
+    if (lonlat == "GLOBAL" & temporal_average != "CLIMATOLOGY") {
+      stop(call. = FALSE,
+           "\nYou have asked for 'GLOBAL' data. However, this is only",
+           "available for 'CLIMATOLOGY'.\n")
+    } else if (lonlat != "GLOBAL") {
+      stop(call. = FALSE,
+           "\nYou have entered an invalid value for `lonlat`. Valid values are",
+           "`GLOBAL` with `CLIMATOLOGY` or a string of lon and lat values.\n")
+    }
+    if (is.character(community)) {
+      community <- toupper(community)
+    }
   }
 
-  # user input checks and formatting -------------------------------------------
-  # see internal_functions.R for these functions
+    # user input checks and formatting -----------------------------------------
+    # see internal_functions.R for these functions
 
-  .check_community(community, pars)
+    .check_community(community, pars)
 
-  dates <- .check_dates(
-    dates,
-    lonlat,
-    temporal_average
-  )
-  pars <- .check_pars(
-    pars,
-    temporal_average,
-    lonlat
-  )
-  lonlat_identifier <- .check_lonlat(
-    lonlat,
-    pars,
-    temporal_average
-  )
+    dates <- .check_dates(
+      dates,
+      lonlat,
+      temporal_average
+    )
+    pars <- .check_pars(
+      pars,
+      temporal_average,
+      lonlat
+    )
+    lonlat_identifier <- .check_lonlat(
+      lonlat,
+      pars
+    )
 
-  # submit query ---------------------------------------------------------------
-  # see internal_functions.R for this function
-  NASA <- .power_query(community,
-    lonlat_identifier,
-    pars,
-    dates,
-    outputList = "CSV"
-  )
-}
+    # submit query -------------------------------------------------------------
+    # see internal_functions.R for this function
+    query_list <- .power_query(community,
+                         lonlat_identifier,
+                         pars,
+                         dates,
+                         outputList = "CSV"
+    )
+    out <- .send_query(.query_list = query_list, .pars = pars)
+    return(out)
+  }