+++ R/internal_functions.R
@@ -208,17 +208,10 @@
       )
     }
 
-    # calculate how many lines to skip in the header to read the CSV from server
-    if (temporal_average == "CLIMATOLOGY") {
-      skip_lines <- length(pars) + 7
-    } else {
-      skip_lines <- length(pars) + 9
-    }
-
     # all good? great. now we format it for the API
     pars <- paste0(pars, collapse = ",")
-    pars <- list(pars, temporal_average, skip_lines)
-    names(pars) <- c("pars", "temporal_average", "skip_lines")
+    pars <- list(pars, temporal_average)
+    names(pars) <- c("pars", "temporal_average")
     return(pars)
   }
 
@@ -234,9 +227,18 @@
 #'
 #' @noRd
 .check_lonlat <-
-  function(lonlat, pars, temporal_average) {
+  function(lonlat, pars) {
     bbox <- NULL
-    if (is.numeric(lonlat) & length(lonlat) == 2) {
+    if (is.character(lonlat) & length(lonlat) == 1) {
+      if (lonlat == "GLOBAL") {
+        identifier <- "Global"
+      } else if (is.character(lonlat)) {
+        stop(
+          call. = FALSE,
+          "\nYou have entered an invalid request for `lonlat`.\n"
+        )
+      }
+    } else if (is.numeric(lonlat) & length(lonlat) == 2) {
       if (lonlat[1] < -180 | lonlat[1] > 180) {
         stop(
           call. = FALSE,
@@ -265,9 +267,9 @@
           "or a 5 x 5 region of 1 degree values, (i.e. 100 points total).\n"
         )
       } else if (any(lonlat[1] < -180 |
-              lonlat[3] < -180 |
-              lonlat[1] > 180 |
-              lonlat[3] > 180)) {
+                     lonlat[3] < -180 |
+                     lonlat[1] > 180 |
+                     lonlat[3] > 180)) {
         stop(
           call. = FALSE,
           "\nPlease check your longitude, `",
@@ -306,8 +308,6 @@
                     lonlat[3],
                     sep = ","
       )
-    } else if (temporal_average == "CLIMATOLOGY") {
-      identifier <- "Global"
     } else {
       stop(
         call. = FALSE,
@@ -349,16 +349,10 @@
                          pars,
                          dates,
                          outputList) {
-  power_url <- # nocov start
-    "https://power.larc.nasa.gov/cgi-bin/v1/DataAccess.py?"
-  client <- crul::HttpClient$new(url = power_url)
+
   user_agent <- "nasapower"
 
-  # check status
-  status <- client$get()
-  status$raise_for_status() # nocov end
-
-  if (lonlat_identifier$identifier == "SinglePoint") {
+  if (lonlat_identifier$identifier == "SinglePoint" & !is.null(dates)) {
     query_list <- list(
       request = "execute",
       identifier = lonlat_identifier$identifier,
@@ -374,13 +368,40 @@
     )
   }
 
-  if (lonlat_identifier$identifier == "Regional") {
+  if (lonlat_identifier$identifier == "SinglePoint" & is.null(dates)) {
+    query_list <- list(
+      request = "execute",
+      identifier = lonlat_identifier$identifier,
+      parameters = I(pars$pars),
+      userCommunity = community,
+      tempAverage = pars$temporal_average,
+      outputList = outputList,
+      lon = lonlat_identifier$lon,
+      lat = lonlat_identifier$lat,
+      user = user_agent
+    )
+  }
+
+  if (lonlat_identifier$identifier == "Regional" & !is.null(dates)) {
     query_list <- list(
       request = "execute",
       identifier = lonlat_identifier$identifier,
       parameters = I(pars$pars),
       startDate = dates[[1]],
       endDate = dates[[2]],
+      userCommunity = community,
+      tempAverage = pars$temporal_average,
+      bbox = I(lonlat_identifier$bbox),
+      outputList = outputList,
+      user = user_agent
+    )
+  }
+
+  if (lonlat_identifier$identifier == "Regional" & is.null(dates)) {
+    query_list <- list(
+      request = "execute",
+      identifier = lonlat_identifier$identifier,
+      parameters = I(pars$pars),
       userCommunity = community,
       tempAverage = pars$temporal_average,
       bbox = I(lonlat_identifier$bbox),
@@ -400,17 +421,33 @@
       user = user_agent
     )
   }
+  return(query_list)
+}
 
-  # send the query
+#' Sends the Query to the API
+#'
+#' @param .query_list A query list created by `.power_query`
+#' @noRd
+#'
+.send_query <- function(.query_list, .pars) {
+  power_url <- # nocov start
+    "https://power.larc.nasa.gov/cgi-bin/v1/DataAccess.py?"
+  client <- crul::HttpClient$new(url = power_url)
+
+  # check status
+  status <- client$get()
+  status$raise_for_status() # nocov end
+
   tryCatch({
-    response <- client$get(query = query_list, retry = 6)
+    response <- client$get(query = .query_list, retry = 6)
     txt <- jsonlite::fromJSON(response$parse("UTF-8"))
     raw_power_data <- file.path(tempdir(), "power_data_file")
   }, # nocov start
   error = function(e) {
     e$message <-
-      paste("\nSomething went wrong with the query, no data were returned.\n",
-            "Please see <https://power.larc.nasa.gov> for potential server issues.\n")
+      paste("\nSomething went wrong with the query, no data were returned.",
+            "Please see <https://power.larc.nasa.gov> for potential",
+            "server issues.\n")
     # Otherwise refers to open.connection
     e$call <- NULL
     stop(e)
@@ -425,41 +462,48 @@
   }
 
   if ("csv" %in% names(txt$output)) {
-    if (outputList == "CSV") {
+    if (.query_list$outputList == "CSV") {
       curl::curl_download(txt$output$csv,
                           destfile = raw_power_data,
                           mode = "wb",
                           quiet = TRUE
       )
 
-      meta <- readLines(
-        raw_power_data,
-        pars$skip_lines
+      power_data <- readLines(
+        raw_power_data
       )
-      meta <- meta[-c(1, pars$skip_lines)] # remove "HEADER ..." lines
+
+      # create meta ojbect
+      meta <- power_data[c(grep("-BEGIN HEADER-",
+                                power_data):grep("-END HEADER-",
+                                                 power_data))]
+      # strip BEGIN/END HEADER lines
+      meta <- meta[-c(1, max(length(meta)))]
+      # replace missing values with NA
       meta <- gsub(
-        pattern = "-99",
+        pattern = "-999",
         replacement = "NA",
         x = meta
       )
 
       power_data <- readr::read_csv(raw_power_data,
                                     col_types = readr::cols(),
-                                    na = "-99",
-                                    skip = pars$skip_lines
+                                    na = "-999",
+                                    skip = length(meta) + 2
       )
 
       # put lon before lat (x, y format)
       power_data <- power_data[, c(2, 1, 3:ncol(power_data))]
 
       # if the temporal average is anything but climatology, add date fields
-      if (pars$temporal_average == "DAILY") {
+      if (.pars$temporal_average == "DAILY") {
         power_data <- .format_dates(power_data)
       }
 
       # add new class
       power_data <- tibble::new_tibble(power_data,
-                                       subclass = "POWER.Info")
+                                       subclass = "POWER.Info",
+                                       nrow = nrow(power_data))
 
       # add attributes for printing df
       attr(power_data, "POWER.Info") <- meta[1]
@@ -496,6 +540,7 @@
 #' @param x POWER.Info object
 #' @param ... ignored
 #' @export
+#' @noRd
 print.POWER.Info <- function(x, ...) {
   if (!is.null(attr(x, "POWER.Info"))) {
     cat(