+++ NEWS.md
@@ -1,3 +1,26 @@
+# nasapower 1.1.0
+
+## Bug fixes
+
+- Fixes bug where .met files were not created properly including where "radn"
+and "rain" col headers were reversed
+
+- Fix `Warning: Must pass a scalar integer as 'nrow' argument to 'new_tibble()'.`
+
+## Major changes
+
+- Change how `GLOBAL` values are requested. This is now specified in `lonlat`
+in conjunction with `temporal_average = CLIMATOLOGY`.
+
+## Minor changes
+
+- Adds example of fetching climate for a single point
+
+- Fixes bug where "CLIMATE" could not be requested for a single point
+
+- Refactor code to split internal functions by functionality and add more
+complete test coverage
+
 # nasapower 1.0.7
 
 ## Minor changes